Change seg label from uint16 to uint32

+++ src/rpad/pybullet_libs/camera.py
@@ -47,11 +47,11 @@ def get_pointcloud(depth, intrinsics) -> npt.NDArray[np.float32]:
 class Render(TypedDict, total=False):
     rgb: npt.NDArray[np.uint8]
     depth: npt.NDArray[np.float32]
-    seg: npt.NDArray[np.uint16]
+    seg: npt.NDArray[np.uint32]
     P_cam: npt.NDArray[np.float32]
     P_world: npt.NDArray[np.float32]
     P_rgb: npt.NDArray[np.uint8]
-    pc_seg: npt.NDArray[np.uint16]
+    pc_seg: npt.NDArray[np.uint32]
     segmap: Dict[int, Tuple[int, int]]
 
 
@@ -169,11 +169,11 @@ class Camera:
         output: Render = {
             "rgb": np.asarray(rgb).astype(np.uint8),
             "depth": np.asarray(depth).astype(np.float32),
-            "seg": np.asarray(seg).astype(np.uint16),
+            "seg": np.asarray(seg).astype(np.uint32),
             "P_cam": np.asarray(P_cam).astype(np.float32),
             "P_world": np.asarray(P_world).astype(np.float32),
             "P_rgb": np.asarray(P_rgb).astype(np.uint8),
-            "pc_seg": np.asarray(pc_seg).astype(np.uint16),
+            "pc_seg": np.asarray(pc_seg).astype(np.uint32),
         }
 
         # Undoing the bitmask so we can get the obj_id, link_index
